Create separate component for perspective switcher.
